requirements.txt

celery==5.4.0

Django==5.1.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7

PyJWT==2.9.0

PyYAML==6.0.2
redis==5.2.0
requests==2.32.3

psycopg2-binary==2.9.10

django-cors-headers

python-decouple==3.8

psycopg2-binary==2.9.10